# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Using a file service as an example",
        "title": "Go Media Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/file/remove/{url}": {
            "delete": {
                "description": "Remove a file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "file"
                ],
                "summary": "Remove",
                "operationId": "url",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File URL",
                        "name": "url",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.removeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        },
        "/file/store": {
            "post": {
                "description": "Save file",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "file"
                ],
                "summary": "Store",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Send file",
                        "name": "request",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.File"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.File": {
            "type": "object",
            "properties": {
                "file_key": {
                    "type": "string",
                    "example": "0b4a4e6e-e502-454f-a8a0-fb7d445e5efa"
                },
                "file_name": {
                    "type": "string",
                    "example": "template.png"
                },
                "file_url": {
                    "type": "string",
                    "example": "https://example.com/images/template.png"
                }
            }
        },
        "v1.removeResponse": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "v1.response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        }
    }
}